implement home screen in search api , implemented mandated field in story , resove dstroy fir nil class in wishlist api erroe

+++ app/controllers/spree/api/v1/wishlists_controller.rb
@@ -27,7 +27,7 @@ module Spree
         end
       rescue Exception => e
         render json:{
-          status: 0,
+          status: "0",
           error: e.message.to_s
         }
       end
@@ -41,30 +41,30 @@ module Spree
             @wishlist  = Spree::Wishlist.new(wishlist_params.merge({user_id: @user.id}))
             if @wishlist.save
               render json: {
-                status: 1 ,
+                status: "1" ,
                 message: "Item added successfully to wishlist"
               }
             else
               render json: {
-                status: 0,
+                status: "0",
                 message: "Something Went Wrong"
               }
             end
           else
             render json: {
-              status: 0,
+              status: "0",
               message: "No variant exist with this id"
             }
           end
         else
           render json: {
-            status: 0,
+            status: "0",
             message: "User not Found"
           }
         end  
       rescue Exception => e
         render json:{
-          status: 0,
+          status: "0",
           error: e.message.to_s
         }
       end 
@@ -73,26 +73,33 @@ module Spree
     def destroy
       begin
         if Spree::Variant.exists?(params[:id])
-          if @user.wishlists.where(variant_id: params[:id]).last.destroy
-            render json: {
-                status: 1 ,
-                message: "Item deleted successfully from wishlist"
-            } 
+          if @user.wishlist_variant_ids.include? params[:id].to_i
+            if @user.wishlists.where(variant_id: params[:id]).last.destroy
+              render json: {
+                  status: "1" ,
+                  message: "Item deleted successfully from wishlist"
+              } 
+            else
+               render json: {
+                  status: "0" ,
+                  message: "Item not deleted successfully from wishlist"
+              }
+            end
           else
-             render json: {
-                status: 0 ,
-                message: "Item not deleted successfully from wishlist"
+            render json:{
+              status: "0",
+              message: "No wishlist exist with this variant id"
             }
           end
         else
           render json: {
-                status: 0,
+                status: "0",
                 message: "No variant exist with this id"
             } 
         end
       rescue Exception => e
         render json:{
-          status: 0,
+          status: "0",
           error: e.message.to_s
         }
       end
